+++ src/app/not-found.tsx
@@ -1,5 +1,5 @@
 import TV404 from "@/app/components/errors/NotFoundView"
-import styles from "@/app/not-found.module.css"; 
+import styles from "@/app/not-found.module.css";
 
 export default function NotFound() {
   return (